update new code

+++ src/main/java/org/example/controller/UserController.java
@@ -1,7 +1,7 @@
 package org.example.controller;
 
-import org.example.entity.CategoryEntity;
 import org.example.entity.UserEntity;
+import org.example.enums.error.UserError;
 import org.example.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -31,6 +31,10 @@ public class UserController {
     @GetMapping("/detail/{id}")
     public String showDetail(@PathVariable("id") long id, Model model) {
         UserEntity user = userService.getById(id);
+        if (user == null) {
+            model.addAttribute("errorMessage", UserError.NOT_FOUND.getMessage());
+            return "error/404";
+        }
         model.addAttribute("user", user);
         return "user/detail";
     }
@@ -51,12 +55,21 @@ public class UserController {
     @GetMapping("/load/{id}")
     public String updateForm(@PathVariable("id") long id, Model model) {
         UserEntity user = userService.getById(id);
+        if (user == null) {
+            model.addAttribute("errorMessage", UserError.NOT_FOUND.getMessage());
+            return "error/404";
+        }
         model.addAttribute("user", user);
         return "user/update";
     }
 
     @PostMapping("/delete/{id}")
-    public String deleteUserById(@PathVariable("id") long id) {
+    public String deleteUserById(@PathVariable("id") long id, Model model) {
+        UserEntity user = userService.getById(id);
+        if (user == null) {
+            model.addAttribute("errorMessage", UserError.NOT_FOUND.getMessage());
+            return "error/404";
+        }
         userService.deleteCategoryById(id);
         return "redirect:/user/list";
     }
